Merge samhays-dev into Dev (#144)
* Remove superfluous DecimalEncoders, Standardize

This commit makes a few changes:
1. Adds the pycommon submodule
2. Removes any DecimEncoder defined here and uses those in pycommon
3. Format some of the code with `black`
4. Add some docstrings
5. Add some type hints

We did have a number of different encoders with different semantics
both used and unused. We can now use the appropriate encoders in
pycommon (see docstrings there). Similarly pycommon introduces
some json encoders helper functions such as dumps_lossy which uses
the encoder that goes from Decimal -> Int. I do wonder when/why this
might be necessary but will discuss with the team later.

* Remove unused files (generator/crud and test)

Some grepping around shows that these files are not in use anywhere.
Code that doesn't exist cannot be a security risk or cause any
confusion.

* Remove another DecimalEncoder and use pycommon

The `decimalencoder` used here was lossy (always converting to int).
Therefore, to keep like-for-like, I've used the `dumps_lossy` helper.
I do wonder if it'd be more appropriate to use an encoder that maintains
floats and ints appropriately.

* Update pycommon

* Begin migration of schemata into their own folder

* Create schema_validation_rules file

This change adds a new file representing the schema validation rules.
This is part of the separation work between what is in `common` and
what belongs in `pycommon`. We want the schema & validation rules to
live in the service but we want the validation _logic_ to live in
common, to ensure consistency across services.

* update pycommon submodule

* update readme to include submodules

* Update pycommon